age: Move SSH identity types into age::ssh module
